Clean up enums
