+++ pybaseball/enums/fangraphs/league.py
@@ -1,0 +1,9 @@
+from enum import Enum, unique
+
+
+@unique
+class FanGraphsLeague(Enum):
+    ALL = 'all'
+    AL  = 'al'
+    FL  = 'fl'
+    NL  = 'nl'
