+++ pybaseball/enums/fangraphs/stat_types.py
@@ -1,0 +1,8 @@
+from enum import Enum, unique
+
+
+@unique
+class FanGraphsStatTypes(Enum):
+    BATTING  = 'bat'
+    FIELDING = 'fld'
+    PITCHING = 'pit'
